Add builds for Windows
As XIVLauncher.Core is a cross-platform project, we should also be providing Windows builds, even if users should typically use the WPF application instead.

These can be used for testing and in rare troubleshooting circumstances where the user is unable to use the WPF application, like when SquirrelSetup constantly fails, AV exceptions aren't working, or if the machine has .Net Framework disabled/removed.