allow override of Steam AppID to FFXIV IDs (#255)
Prior to ebc8bcebc67050cf252b37cb3076b5c14e3c661d it was possible to set
Steam AppID to Free Trial by selecting the global "free trial" option.
The new per-account setting removed the if/else branch that made that
possible, and now Retail ID is always prioritized. This creates an
unintuitive experience for non-Steam players (on SteamDeck, for example)
who have a family member owning a Steam edition of FFXIV, where the game
will always pick Retail AppID instead of Free Trial, regardless if the
user has Retail installed or not. That's not only unintuitive but it
also prevents Steam edition owning family member from launching the game
at the same time.

The current workaround for this is to set XL_APPID to a non-FFXIV ID,
which is rather hacky and means the player is forced to set FFXIV
controller layout to whatever game was picked.

To improve the experience, this patch changes the AppID priority order
logic, so that SteamAppID or XL_APPID environment variables can be used
to give Retail/FT IDs a higher priority when attempting to select which
AppID to use.